% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{add_logo}
\alias{add_logo}
\title{Add a logo and styled borders to a ggplot visualization}
\usage{
add_logo(
  plot,
  corner_radius = 0.02,
  border_color = light_blue_color,
  border_size = 2,
  logo_path = system.file("extdata", "logo.png", package = "tariffTheme"),
  logo_position = "top right",
  logo_width = 0.15
)
}
\arguments{
\item{plot}{A ggplot object to enhance with logo and borders}

\item{corner_radius}{The radius of the rounded corners, from 0 to 1}

\item{border_color}{The color of the border, defaults to light blue}

\item{border_size}{The thickness of the border}

\item{logo_path}{Path to the logo image file, defaults to the package's logo}

\item{logo_position}{Position of the logo (one of "top right", "top left", "bottom right", "bottom left")}

\item{logo_width}{Proportional width of the logo}
}
\value{
A finalized plot with logo and borders
}
\description{
This function adds the tariff theme logo and styled borders to a ggplot visualization,
creating a finished, branded plot ready for presentation or publication.
}
\examples{
library(ggplot2)
p <- ggplot(country_tariffs, aes(x = country, y = reciprocal_tariff_adjusted)) +
  geom_bar(stat = "identity", fill = yellow_color) +
  labs(title = "Reciprocal Tariffs") +
  theme_tariff()
add_logo(p)
}
