% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-country_tariffs.R, R/data.R
\docType{data}
\name{country_tariffs}
\alias{country_tariffs}
\title{Country Tariffs Dataset}
\format{
A data.table with 57 rows and 2 variables:
\describe{
\item{country}{Country name}
\item{reciprocal_tariff_adjusted}{Reciprocal tariff rate as a percentage}
}

A data.table with 57 rows and 2 variables:
\describe{
\item{country}{Country name}
\item{reciprocal_tariff_adjusted}{Reciprocal tariff rate as a percentage}
}
}
\source{
Synthetic data for illustration purposes

Synthetic data for illustration purposes
}
\usage{
country_tariffs

country_tariffs
}
\description{
A dataset containing reciprocal tariff rates for various countries.
This dataset provides tariff information that can be used for
demonstrations and examples with the tariffTheme package.

A dataset containing reciprocal tariff rates for various countries.
This dataset provides tariff information that can be used for
demonstrations and examples with the tariffTheme package.
}
\examples{
data(country_tariffs)
head(country_tariffs)
}
\keyword{datasets}
