% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_tariff}
\alias{scale_color_tariff}
\title{Custom color scale for tariff theme}
\usage{
scale_color_tariff(...)
}
\arguments{
\item{...}{Additional color values to append to the default palette}
}
\value{
A ggplot2 scale object
}
\description{
A function that provides a color scale using the theme's custom colors
}
\examples{
library(ggplot2)
# Use a smaller subset of data for the example
ggplot(head(country_tariffs, 10), aes(x = country, y = reciprocal_tariff_adjusted, fill = country)) +
  geom_bar(stat = "identity") +
  scale_fill_tariff() +
  theme_tariff(use_custom_font = FALSE)
}
