% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{theme_colors}
\alias{theme_colors}
\alias{navy_color}
\alias{light_blue_color}
\alias{yellow_color}
\title{Theme colors for tariff visualizations}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
navy_color

light_blue_color

yellow_color
}
\description{
A set of color variables used consistently across the tariff theme.
These colors form the core brand palette for all tariff visualizations.
}
\keyword{datasets}
